Hopefully fix auth flow
